+++ projects/cm_mcu/MonitorTask.h
@@ -39,10 +39,10 @@ struct MonitorTaskArgs_t {
   struct dev_i2c_addr_t *devices;      // list of devices to query
   int n_devices;                       // number of devices
   struct pm_command_t *commands;       // list of commands
-  const int n_commands;                // number of commands
+  const uint8_t n_commands;                // number of commands
   float *pm_values;                    // place to store results
   const int n_values;                  // number of results
-  const int n_pages;                   // number of pages to loop over
+  const uint8_t n_pages;                   // number of pages to loop over
   tSMBus *smbus;                       // pointer to I2C controller
   volatile tSMBusStatus *smbus_status; // pointer to I2C status
   volatile TickType_t updateTick;      // last update time, in ticks
